pyproject.toml
[project]
name = "diagnoseme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.8.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "tomli-w>=1.2.0",
    "tortoise-orm>=1.1.7",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "coverage>=7.13.5",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 120

[tool.mypy]
plugins = [
    "pydantic.mypy",
]
python_version = 3.14
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
